fix: handle limits and offsets after filtering  (#281)
I took a look at the source code today because I had some time, and I think that this occurs because in both manager_sql.go and manager.go get the members based on the limit/offset first, before applying the filter in ListByQuery.

The gist of this PR is that I inspect the query parameters first to see if filters are being applied. If filters are applied, it will retrieve all of the roles/policies, instead of simply the limit/offset specified. Once in memory, I calculate the pagination again in ListByQuery to shorten the array.

The retrieval from the database (or in-memory datastore) is completely separate from the filtering logic. 
As such, in order to handle this, I needed to retrieve all the possible data, then run the filter on it. After running the filter on it, then I could shorten it by the pagination, in order to ensure that it did not return an empty array.

In addition, I also did a minor refactor on the tests, instead of duplicating all the Roles/Policies in `filter_helper_tests.go`, I simply referred to them from the already declared variable `rolReq` or `polReq` to ensure that updating the role or policy in one place would change across all the rests of the tests. In addition, I also changed the parameters from a `map[string][]string` into a structure, allowing me to specify the offset and limit I wanted to test for each case.

Fixes #280
Fixes #216