refactor: make subject sets and subject IDs unambiguous (#729)
BREAKING CHANGES: This patch changes the payload of the REST API. The gRPC API is **not** affected. The parameter `subject` was previously an encoded string. With this change clients have to explicitly use either `subject_id` or (`subject_set.namespace` and `subject_set.object` and `subject_set.relation`). The same is true for REST responses returned by Keto. An error with a hint will be returned if `subject` is still used.