fix: remove pagination-related query elements on LIMIT 1
This patch removes query components which cause CockroachDB to perform a full-range query across all leaseholders for LIMIT 1 queries. Essentially, CockroachDB will query all potential nodes which might hold data when ORDER BY is used, or when the LIMIT indicates that there are rows missing.

This can drop the query latency to a few ms versus hundreds of ms on a CockroachDB `REGIONAL BY ROW` table, without impact on the business logic.

Closes #1167